Take more care about stack in Reset_Handler
In some builds the compiler was pushing registers to the stack on entry
to Reset_Handler, rendering the setting of MSP there rather pointless.

Noticed because Ozone was loading an incorrect value due to incorrectly
guessing the vector table was in the ITCM (lowest-addressed read-only
area? Not clear why it wasn't assuming that about the DTCM before).