Adjust alif_kws use case loop
For live audio use cases it makes more sense for the UseCaseHandler to
just keep running - doing multiple inferences on a stream of audio as
the original use cases do for a single audio file. For now, just let
it loop indefinitely until error - in future we could give it a time
limit.

This is different from live video use cases, where each frame of
video is a separate operation equivalent to one image file.

In particular this avoids the CPU-intensive setup of KwsPreProcess and
KwsPostProcess for every stride.

Couple of other minor adjustments - sample rate can now be configured in
CMake, and the audio record buffer is aligned, just in case there's a
DMA cache coherency issue. (Can't be with current memory maps, because
it's in DTCM).