Invalidate cache once camera capture finishes
Do the Right Thing and invalidate /after/ the camera has finished,
rather than while it's doing it. Yes, this is a bit slower, and no,
we haven't seen the theoretical problem of a prefetch filling
the cache while it was running, but the speed hit is negligible
compared to the actual processing, especially now we've reduced
the size of the buffer!