Eliminate unneeded Tensorflow and CMSIS-NN code
When using the Ethos-U, we don't need to include the software
implementation of operators that it is handling. Omit them by default
for the specific use case models.

Thie is controlled via a ETHOS_U_NPU_ASSUMED option that defaults to
true. This can be turned off if you want to run the non-Vela version of
a model.

The inference_runner use case continues to register all operators, as a
general-purpose use case.