MLECO-5482: Enabling support for CMSIS-6
Adding support to build against CMSIS-6 sources. CMake configuration
parameter `CMSIS_VER` can be set to use CMSIS versions 5 or 6; the
default is kept at 5 with the aim to update it to 6 once it has been
tested sufficiently.

This patch renames the old submodule from `cmsis` to `cmsis-5` and
adds a new submodule `cmsis-6`. Another dependency downloaded during
CMake configuration is one file from
https://git.mlplatform.org/ml/ethos-u/ethos-u-core-software.git. If
we need more files from this repo, it may be worth adding this as a
submodule too.

Change-Id: I0239f253c1c69f9adab33900b08323cdcf7b67d6
Signed-off-by: Kshitij Sisodia <kshitij.sisodia@arm.com>