MLECO-5339: Adding HAL camera and audio interfaces
Adding a HAL camera and audio interface to allow these as standard
APIs used across use cases. This exercise is primarily to pave the
way for adding Virtual Streaming Interface (VSI) support for Arm
Fixed Virtual Platforms (FVPs) in future work. However, this also
makes the application more portable and easier to add support for
physical (camera or microphone) devices, should we choose to do so.

This patch is big as it cuts across use cases and HAL layers with
platform support. The implementation of baked-in sample images and
audio data is quite old in this repository. To wire things up via
HAL for something which was only application level required touching
quite a few files.

General outline of changes:
--------------------------
 * File generation part has been changed so it generates a self
   contained set of C files in a separate directory that provide
   strongly linked functions that HAL camera or audio interface
   relies on.
 * Two new "components" (camera and audio) under HAL. They have
   weak implementations of functions that should be overridden by
   application sources. The reason for doing it this way is different
   files are generated for different use cases but the there is only
   one target we need under HAL - it is built only once but different
   application sources can override and use this module differently.
 * All the use cases changed to move away from using the generated
   files' functions directly. Instead the static baked-in data is now
   accessed via the HAL provided API's.
 * Tests for all use cases changed for same reason as above.

Byproduct of above changes:
--------------------------
 * Apps don't need to be interactive (user input). The application
   level cannot have knowledge of how many baked in samples we have
   or the filenames. This knowledge only lives below HAL.
 * Tests which were validating inferences by index have been removed.

Change-Id: I09964ce87ea5cea5700d141eef199d3d2be6686c
Signed-off-by: Kshitij Sisodia <kshitij.sisodia@arm.com>